Fetch live events from Eventful.
